+++ lib/fog/azurerm/models/compute/servers.rb
@@ -1,6 +1,3 @@
-require 'fog/core/collection'
-require 'fog/azurerm/models/compute/server'
-
 module Fog
   module Compute
     class AzureRM
@@ -8,21 +5,21 @@ module Fog
       # for Server.
       class Servers < Fog::Collection
         attribute :resource_group
-        model Fog::Compute::AzureRM::Server
+        model Server
 
         def all
           requires :resource_group
           virtual_machines = []
           service.list_virtual_machines(resource_group).each do |vm|
-            virtual_machines << Fog::Compute::AzureRM::Server.parse(vm)
+            virtual_machines << Server.parse(vm)
           end
           load(virtual_machines)
         end
 
         def get(resource_group_name, virtual_machine_name)
-          storage_account = service.get_virtual_machine(resource_group_name, virtual_machine_name)
-          storage_account_obj = Fog::Compute::AzureRM::Server.new(service: service)
-          storage_account_obj.merge_attributes(Fog::Compute::AzureRM::Server.parse(storage_account))
+          virtual_machine = service.get_virtual_machine(resource_group_name, virtual_machine_name)
+          virtual_machine_obj = Server.new(service: service)
+          virtual_machine_obj.merge_attributes(Server.parse(virtual_machine))
         end
       end
     end